Add files via upload
Through this visit, it was a good opportunity to let me know the various platforms of UTM such as UTMSmart, elearning and MyUtm Portal, and how these all platforms operate on the backend through the visit to UTM Digital. This visit also improve my understanding about TIS course and IT field.